+++ CHANGES.rst
@@ -9,6 +9,13 @@
 - Add a BTree "family" object to ``nti.zodb.btrees`` that uses larger
   bucket sizes. See `issue 8 <https://github.com/NextThought/nti.zodb/issues/8>`_.
 
+- All numeric minmax objects implement the same interface, providing
+  the ``increment`` method. See `issue 7
+  <https://github.com/NextThought/nti.zodb/issues/7>`_.
+
+- The merging counter does the right thing when reset to zero by two
+  conflicting transactions. See `issue 6
+  <https://github.com/NextThought/nti.zodb/issues/6>`_.
 
 1.1.0 (2020-07-15)
 ==================